(maint) update docs to use charset and add clear_env_var
this updates the README and tests to encourage the user to use the same locale indentifier for both methods (ja_JP.utf8) instead of for just change_locale_on. this also adds the use of the clear_env_var method in change_locale_on to repair the inability to change the locale more than once. now, the locale is cleared and reset each time this method is called.